checking 2D error behavior
doing this on errmap branch for now (and for fun).  Added guess and error tracking to .  Also ipynb to start to look into behavior.  For the case considered, it's clear the root finder is very inefficient, stuck on a marginally good solution and not improving for the last 20 guesses (most of them).