good progress on error mapping!
As figure shows with dots the rootfinders guesses are settling for bad solutions even though the solution space is smooth.  Given that finding the true root (once this close) is as simple as following the intersecting lines, it should be easy to write a root finder (or root refiner) that does much better relatively quickly.