noise in 2D error derivatives (Jacobian) at small stepsizes
The noise shown in the Jac_vs_stepsize fig is NOT due to roundoff error, as argued in the notebook.  The shooting method must have some noise in the error generation at very low levels.  It's not clear that this is the main issue.  There is clearly a regime where the Jacobian is resolved, so next step is to look if the well behaved jacobian points in the right direction.