modified analytic comparisons
I did this as an addition not a subtraction.  So feel free to clean up the plots_poster notebook after you look at my changes, which I'm hoping I can merge.  For the analytic model I added a (numerical) computation of the radiative mass for a fairer comparison.  Now the agreement at early times is good at least.  I did change so that L_cb depends only on core mass which is what the analytic calculation actually does. This made no visible difference.