reshuffling of sections in outline
All methods in section 2, which should gradually build from the general to the specific.  Ana, if you're going to start writing here, just keep it as basic and factual as possible and not too wordy.  I.e. just get the basic equations and a few connecting words to the page.  The flowery language if anywhere belongs elsewhere and later modifications are possible.