finally a true root found!
Still only for a single step from a static solution, but much better errors that before with a simple approach.  The basic jacobian works better than the advanced method.  The difference may avoiding the numerical noise in the Jacobian at high resolution.  but it may also be that a stupid method does better than the smart one because the error space is so smooth.  One small step.