s2.root2D modified to add ezroot option
Also modified ezroot to return a namedtuple that matches the form of the output from the scipy rootfinders, e.g. sol.x for solution.  Not as much info though.  Also some improvements to keyword passing with ** unpacking.  That could still be done more to simplify code with shorter argument lists (and also reduce probability of typos).